# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Challenger API",
        "description": "Challenger API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1",
            "description": "challenger-api server"
        }
    ],
    "paths": {
        "/api/v1/locations": {
            "get": {
                "tags": [
                    "locations"
                ],
                "summary": "Get all locations",
                "description": "Gets a list of all locations and allows filtering with the name parameter",
                "operationId": "9bbe66ccd9a8de4d94275b1f6c4be2b9",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter locations by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of all locations"
                    },
                    "404": {
                        "description": "No location found with the specified name."
                    },
                    "500": {
                        "description": "An error occurred in the listing of the locations."
                    }
                }
            }
        },
        "/v1/location": {
            "post": {
                "tags": [
                    "locations"
                ],
                "summary": "Create location",
                "description": "Create a new location",
                "operationId": "6963a88a397f84cc375e96c6beeb7a67",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Location Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "query",
                        "description": "Location slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Location city",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "description": "Location state",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Location successfully created."
                    },
                    "422": {
                        "description": "The slug has already been taken or param is required"
                    },
                    "500": {
                        "description": "An error occurred in the creation of the location."
                    }
                }
            }
        },
        "/api/v1/locations/{id}": {
            "get": {
                "tags": [
                    "locations"
                ],
                "summary": "Get a location by id",
                "description": "Get a location by id parameter",
                "operationId": "6bc92e1682723116be7ea8ff826b9f4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the location",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location info list"
                    },
                    "404": {
                        "description": "No location found with the specified id."
                    },
                    "500": {
                        "description": "An error occurred in the listing of the locations."
                    }
                }
            },
            "delete": {
                "tags": [
                    "locations"
                ],
                "summary": "Delete a location by id",
                "description": "Delete a location by id parameter",
                "operationId": "d69677f30e8f588247d039bb36d50082",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the location",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location successfully deleted."
                    },
                    "404": {
                        "description": "No location found with the specified id."
                    },
                    "500": {
                        "description": "An error occurred in the deleted of the location."
                    }
                }
            }
        },
        "/v1/location/{id}": {
            "put": {
                "tags": [
                    "locations"
                ],
                "summary": "Update location by id",
                "description": "Update a location by id",
                "operationId": "fb3ecfc4ecdea8ef10a7e56359f46539",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the location",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Location Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "query",
                        "description": "Location slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Location city",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "description": "Location state",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location successfully updated."
                    },
                    "404": {
                        "description": "No location found with the specified id."
                    },
                    "422": {
                        "description": "The slug has already been taken or param is required"
                    },
                    "500": {
                        "description": "An error occurred in the update of the location."
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "locations",
            "description": "locations"
        }
    ]
}